fix: move all pad and pins to the asic folder fix: do not always generate asic files when doing make mcu-gen fix: minor fixes and remove duplicates
