+++ docs/source/How_to/ImplementASIC.md
@@ -76,7 +76,7 @@ make openroad-sky130
 ### Overview
 A Librelane integration of `x-heep` can perform a full ASIC open-source flow, thus not relying on any proprietary tools.
 
-The whole integration is based on the [ihp-sg13g2-librelane-template](https://github.com/IHP-GmbH/ihp-sg13g2-librelane-template). In particular, the pdn config files and sdc files are the one found in this repo, templated and adapted to `x-heep`.
+The whole integration is based on the [ihp-sg13g2-librelane-template](https://github.com/IHP-GmbH/ihp-sg13g2-librelane-template). In particular, the PDN config files and SDC files are the one found in this repo, templated and adapted to `x-heep`.
 
 There are currently two Librelane implementations, targeting the IHP-SG13G2 PDK:
 - The `asic-librelane-chip-ihp` target whose goal is to implement the microcontroller itself as a full chip, without the external extensions (no CVXIF, no external peripherals, ...). This flow is based on the Librelane [`Chip`](https://librelane.readthedocs.io/en/latest/reference/flows.html#chip) flow. It is aimed at people wanting to use `x-heep` as a standalone microcontroller, without extending it beyond the already implemented peripherals.
@@ -87,6 +87,15 @@ In order to run both flows, you will need to install Librelane as explained in i
 Both flows were tested by hand using `questasim` with the `example_gpio_toggle` code. **You need to simulate the netlist `.nl.v` files generated by Librelane to be sure that there are no issues.**
 
 ### Run commands
+First you need to generate all the template files with the corresponding config and pad files:
+```
+make mcu-gen TARGET=asic_ihp                                \
+             X_HEEP_CFG=configs/python_unsupported.hjson    \
+             PYTHON_X_HEEP_CFG=configs/general_librelane.py \
+             PADS_CFG=configs/pad_cfg_librelane.py
+```
+Note that you can change by hand the generated files inside `hw/asic/ihp-sg13g2` to fit you own design and specify SRAMs placement, SDC constraints, PDN,...
+
 Two targets are present in the Makefile, to run one of the two flow, simply run the corresponding command:
 ```
 make librelane-setup-chip-ihp
@@ -113,3 +122,4 @@ You can find here some of the issues encountered when running the Librelane flow
     - The Metal 1 off-grid errors: These are errors inside the SRAMs macros, and [are known](https://github.com/IHP-GmbH/ihp-sg13g2-librelane-template/issues/16). They should not be an issue.
     - For the chip target, there are some issues with the space between the IO pads and the bondpads if using the `dev` branch. This issue is not present when using the `main` branch of Librelane. However, at the time of writing this, it is needed to use the `dev` branch because [this commit](https://github.com/librelane/librelane/commit/0d7245d99a9eacb46f1ffabe6843747ee5836bfa) is not yet merged to `main`.
 - There are some antenna errors, however these are only reported by the `OpenROAD.CheckAntennas` step, and not by the later run `KLayout.Antenna` step. According to Librelane maintainers, the step that is more trustable is the KLayout one, as it performs the checks on the finished GDS, while the OpenROAD one does the checks on an abstract view of the design.
+- The frequency is really low and does not passes STA: this might be due to the incorrect setting of SDC constraints.
