+++ hw/asic/ihp-sg13g2/librelane_chip_config.yaml.tpl
@@ -1,3 +1,7 @@
+<%
+    if impl_target != "asic_ihp":
+        return STOP_RENDERING
+%>
 <%!
     from pads.pin import Input, Output, Inout
     from pads.pad import Pad
