+++ hw/asic/ihp-sg13g2/librelane_sram_placement.cfg.tpl
@@ -1,3 +1,7 @@
+<%
+    if impl_target != "asic_ihp":
+        return STOP_RENDERING
+%>
 <%!
     from memory_ss.memory_ss import MemorySS
     from memory_ss.ram_bank import Bank
