+++ hw/asic/ihp-sg13g2/rtl/cells/pad_cell_inout_ihp_sg13g2.sv
@@ -1,6 +1,12 @@
 module pad_cell_inout #(
     parameter PADATTR = 0
 ) (
+    `ifdef USE_POWER_PINS
+    inout logic iovdd,
+    inout logic iovss,
+    inout logic vdd,
+    inout logic vss,
+    `endif
     input logic pad_in_i,
     input logic pad_oe_i,
     output logic pad_out_o,
@@ -10,12 +16,12 @@ module pad_cell_inout #(
 
     (* keep *)
     sg13g2_IOPadInOut4mA pad_cell_inout (
-        // `ifdef USE_POWER_PINS
-        // .iovdd(),
-        // .iovss(),
-        // .vdd(),
-        // .vss(),
-        // `endif
+        `ifdef USE_POWER_PINS
+        .iovdd,
+        .iovss,
+        .vdd,
+        .vss,
+        `endif
         .pad(pad_io),
         .c2p(pad_in_i),
         .c2p_en(pad_oe_i),
