+++ hw/asic/ihp-sg13g2/rtl/cells/pad_cell_input_ihp_sg13g2.sv
@@ -1,6 +1,12 @@
 module pad_cell_input #(
     parameter PADATTR = 0
 ) (
+    `ifdef USE_POWER_PINS
+    inout logic iovdd,
+    inout logic iovss,
+    inout logic vdd,
+    inout logic vss,
+    `endif
     input logic pad_in_i,
     input logic pad_oe_i,
     output logic pad_out_o,
@@ -10,12 +16,12 @@ module pad_cell_input #(
 
     (* keep *)
     sg13g2_IOPadIn pad_cell_input (
-        // `ifdef USE_POWER_PINS
-        // .iovdd(),
-        // .iovss(),
-        // .vdd(),
-        // .vss(),
-        // `endif
+        `ifdef USE_POWER_PINS
+        .iovdd,
+        .iovss,
+        .vdd,
+        .vss,
+        `endif
         .pad(pad_io),
         .p2c(pad_out_o)
     );
