+++ hw/asic/ihp-sg13g2/technology_prim_ihp-sg13g2.core
@@ -21,13 +21,16 @@ filesets:
       - rtl/cells/pad_cell_power_ihp_sg13g2.sv
     file_type: systemVerilogSource
 
-  files_chip:
+  files_common:
     depend:
       - x-heep:ip:pad_control
     files:
+      - librelane_sram_placement.cfg : { file_type: macroPlacement }
+      - librelane_pdn_cfg.tcl        : { file_type: pdn }
+
+  files_chip:
+    files:
       - librelane_chip_config.yaml              : { file_type: librelaneConfig }
-      - librelane_chip_macro_placement.cfg      : { file_type: macroPlacement }
-      - librelane_chip_pdn_cfg.tcl              : { file_type: pdn }
       - librelane_chip_fallback_pnr_signoff.sdc : { file_type: SDC }
       - rtl/pad_ring.sv
       - rtl/ihp_sg13g2_asic_chip_x_heep_system_wrapper.sv
@@ -36,11 +39,8 @@ filesets:
   files_macro:
     depend:
       - x-heep:ip:cv32e40px # To make sure the CVXIF signals will be included before the other files
-      - x-heep:ip:pad_control
     files:
       - librelane_macro_config.yaml              : { file_type: librelaneConfig }
-      - librelane_macro_macro_placement.cfg      : { file_type: macroPlacement }
-      - librelane_macro_pdn_cfg.tcl              : { file_type: pdn }
       - librelane_macro_pin_order.cfg            : { file_type: pinOrder }
       - librelane_macro_fallback_pnr_signoff.sdc : { file_type: SDC }
       - rtl/ihp_sg13g2_asic_macro_x_heep_system_wrapper.sv
@@ -52,7 +52,9 @@ targets:
     filesets:
       - target_asic_librelane_chip_ihp ? (pdk_std_cells)
       - target_asic_librelane_chip_ihp ? (pdk_io_cells)
+      - target_asic_librelane_chip_ihp ? (files_common)
       - target_asic_librelane_chip_ihp ? (files_chip)
 
       - target_asic_librelane_macro_ihp ? (pdk_std_cells)
+      - target_asic_librelane_macro_ihp ? (files_common)
       - target_asic_librelane_macro_ihp ? (files_macro)
